renders: use floats in viewport

+++ src/client/refresh/gl4/gl4_main.c
@@ -1339,10 +1339,10 @@ SetupGL(void)
 	int x, x2, y2, y, w, h;
 
 	/* set up viewport */
-	x = floor(gl4_newrefdef.x * vid.width / vid.width);
-	x2 = ceil((gl4_newrefdef.x + gl4_newrefdef.width) * vid.width / vid.width);
-	y = floor(vid.height - gl4_newrefdef.y * vid.height / vid.height);
-	y2 = ceil(vid.height - (gl4_newrefdef.y + gl4_newrefdef.height) * vid.height / vid.height);
+	x = floor(gl4_newrefdef.x * vid.width / (float)vid.width);
+	x2 = ceil((gl4_newrefdef.x + gl4_newrefdef.width) * vid.width / (float)vid.width);
+	y = floor(vid.height - gl4_newrefdef.y * vid.height / (float)vid.height);
+	y2 = ceil(vid.height - (gl4_newrefdef.y + gl4_newrefdef.height) * vid.height / (float)vid.height);
 
 	w = x2 - x;
 	h = y - y2;
